Исправлены пердупреждения из парсера, связанные с сравнением переменных с разными знаками

+++ KAMEditor/gpr/gcode_program.h
@@ -381,7 +381,7 @@ namespace gpr {
     int num_blocks() const { return blocks.size(); }
 
     block get_block(const int i) {
-      assert(i < blocks.size());
+      assert((size_t) i < blocks.size());
       return blocks[i];
     }
